debug: Day8 rust was having a u32 wrapping issue!
- Day8P1 success now. Ugh so annoying! I'm spoiled by python
- Issue was in n_closest_pairs_of_boxes where d calc would
  exceed u32 limit and therefore wrap around to a small positive
  number and mess up the order of the joined boxes

+++ src/lib.rs
@@ -74,12 +74,12 @@ fn day7_part2(data_path: &str) -> PyResult<u128> {
 }
 
 #[pyfunction]
-fn day8_part1(data_path: &str) -> PyResult<u32> {
+fn day8_part1(data_path: &str) -> PyResult<u64> {
     Ok(days::day8::part1(Path::new(data_path)))
 }
 
 #[pyfunction]
-fn day8_part2(data_path: &str) -> PyResult<u32> {
+fn day8_part2(data_path: &str) -> PyResult<u64> {
     Ok(days::day8::part2(Path::new(data_path)))
 }
